- FormaEntregaServiceTest.java -> Alterando ordem de expected e atual; - UnidadeMedida.java -> Adicionando construtores; - UnidadeMedidaController.java -> Alterando parametros de dto para entity e retorno do metodo deletar para String; - UnidadeMedidaService.java -> Adicionando verificações e retorno do metodo deletar; - UnidadeMedidaServiceTest.java -> Adicionada a classe e 10 casos de teste;

+++ src/main/java/school/sptech/apimiseenplace/service/UnidadeMedidaService.java
@@ -2,9 +2,8 @@ package school.sptech.apimiseenplace.service;
 
 import lombok.RequiredArgsConstructor;
 import org.springframework.stereotype.Service;
-import school.sptech.apimiseenplace.dto.unidadeMedida.UnidadeMedidaCriacaoDto;
-import school.sptech.apimiseenplace.dto.unidadeMedida.UnidadeMedidaMapper;
 import school.sptech.apimiseenplace.entity.UnidadeMedida;
+import school.sptech.apimiseenplace.exception.BadRequestException;
 import school.sptech.apimiseenplace.exception.NaoEncontradoException;
 import school.sptech.apimiseenplace.repository.UnidadeMedidaRepository;
 
@@ -17,26 +16,32 @@ public class UnidadeMedidaService {
 
     public UnidadeMedida buscarPorId(Integer id) {
         return unidadeMedidaRepository.findById(id).orElseThrow(
-                () -> new NaoEncontradoException("Unidade de Medida")
+                () -> new NaoEncontradoException("Unidade Medida")
         );
     }
 
-    public UnidadeMedida criar(UnidadeMedidaCriacaoDto unidadeMedidaCriacaoDto) {
-        UnidadeMedida unidadeMedida = UnidadeMedidaMapper.toEntity(unidadeMedidaCriacaoDto);
-        return unidadeMedidaRepository.save(unidadeMedida);
+    public UnidadeMedida criar(UnidadeMedida unidadeMedidaCriacaoDto) {
+        if (unidadeMedidaCriacaoDto == null) throw new BadRequestException("Unidade Medida");
+        return unidadeMedidaRepository.save(unidadeMedidaCriacaoDto);
     }
 
     public List<UnidadeMedida> listar() {
         return unidadeMedidaRepository.findAll();
     }
 
-    public UnidadeMedida atualizar(Integer id, UnidadeMedidaCriacaoDto unidadeMedidaAtualizacao) {
+    public UnidadeMedida atualizar(Integer id, UnidadeMedida unidadeMedidaAtualizacao) {
+        if (unidadeMedidaAtualizacao == null) throw new BadRequestException("Unidade Medida");
+        if (id == null) throw new BadRequestException("Id Unidade Medida");
         UnidadeMedida unidadeMedida = buscarPorId(id);
         unidadeMedida.setUnidadeMedida(unidadeMedidaAtualizacao.getUnidadeMedida());
         return unidadeMedidaRepository.save(unidadeMedida);
     }
 
-    public void deletar(Integer id) {
+    public String deletar(Integer id) {
+        if(!unidadeMedidaRepository.existsById(id)){
+            throw new NaoEncontradoException("Unidade de Medida");
+        }
         unidadeMedidaRepository.deleteById(id);
+        return "Unidade Medida deletada com sucesso!";
     }
 }
